Update fail.js
I'm not always looking at my terminal when grunt decides to go and do its thing. So when I see a failure message, I can't be completely sure which change killed it. Or when the build was last triggered.
Adding a "last failure" time stamp to the output helps a lot.